Use querystringify rather than url-parse
The code imported the url-parse package as follows:

    import * as parse from 'url-parse';

and then used `parse()` as a function. This is not valid per spec,
since "parse" is just a namespace, and makes it impossible to
build this code with module bundlers that follow the spec, such
as rollup.

The simplest solution seems to be to use the querystringify
package directly: the url-parse package uses querystringify
to parse query strings in any case.

As a happy side-effect of this change, the compiled bundle
`dscc.min.js` is significantly smaller.